Fix build by disabling audio features
