+++ src/constants.rs
@@ -1,0 +1,14 @@
+pub const WINDOW_WIDTH: f32 = 800.0;
+pub const WINDOW_HEIGHT: f32 = 600.0;
+pub const PLAYER_SPEED: f32 = 100.0;
+pub const JUMP_FORCE: f32 = 300.0;
+pub const GRAVITY: f32 = -800.0;
+pub const SPRITE_SIZE: f32 = 32.0;
+pub const CHEMICAL_PROJECTILE_SPEED: f32 = PLAYER_SPEED * 2.0;
+pub const CHEMICAL_PROJECTILE_SIZE: f32 = 16.0;
+pub const CHEMICAL_PROJECTILE_COOLDOWN: f32 = 0.6;
+pub const CHEMICAL_PROJECTILE_ROTATION_SPEED: f32 = 2.0;
+pub const ENEMY_SPEED: f32 = 60.0;
+pub const ENEMY_SIZE: f32 = 32.0;
+pub const WAVE_AMPLITUDE: f32 = 50.0;
+pub const WAVE_FREQUENCY: f32 = 2.0;
